requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.2
colorama==0.4.6
cryptography==46.0.7
et_xmlfile==2.0.0
fastapi==0.136.0
google-api-core==2.30.3
google-api-python-client==2.194.0
google-auth==2.49.2
google-auth-httplib2==0.3.1
google-auth-oauthlib==1.3.1
googleapis-common-protos==1.74.0
greenlet==3.4.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.2
idna==3.11
openpyxl==3.1.5
pandas==3.0.2
proto-plus==1.27.2
protobuf==7.34.1
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.13.3
pydantic_core==2.46.3
pyparsing==3.3.2
python-dotenv==1.2.2
python-telegram-bot==22.7
six==1.17.0
SQLAlchemy==2.0.49
starlette==1.0.0
typing-inspection==0.4.2
uritemplate==4.2.0
urllib3==2.6.3
uvicorn==0.45.0
